// src/data/path.enum.ts
export enum Path {
    AUTH = '/auth',
    LOGIN = '/login',
    DASHBOARD = '/dashboard',
    USERS = '/users',
    USERS_DETAILS = '/users/:id',
    PARKINGS = '/parkings',
    PARKING_DETAILS = '/parkings/:id',
    BOOKINGS = '/bookings',
    BOOKING_DETAILS = '/bookings/:id',
    PROPERTY_INTEREST = '/property-interests',
    PROPERTY_INTEREST_DETAILS = '/property-interests/:id',
    CUSTOMER_SUPPORT = "/customer-support",
    PRIVACY_POLICY = "/privacy-policy",
    TERMS_AND_CONDITIONS = "/terms-and-conditions",

}